Fix segfault on error
In newer versions of grpc a new field `error_string` was introduced
somewhere deep inside `grpc_op` structure. If it's not
initialized (i.e. has garbage), grpc can try to write to arbitrary
memory location. This error reporting can be disabled by setting
`error_string` to null.

Ideally, this should be used in a meaningful way instead - i.e. by
properly exposing this to perl side. But this is for later.