Use blocking shutdown call to prevent race if fork is used
Replace the call to `grpc_shutdown()` with a call to `grpc_shutdown_blocking()`
when the gRPC library implements it (versions >= 1.20). This keeps
`Grpc::XS::destroy()` synchronous, allowing it to be safely called before
fork.

Fixes #22.

Update test `17_fork_friendliness.t` to further exercise init, destroy, and
fork and optionally repeat the test in a loop (to catch races).